# api/swagger.yaml
...
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "FC Api"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "agnetiuslee@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "petstore.swagger.io"
basePath: "/ronaldSengkey/fitClub/api/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /coach/{param}/{token}:
    put:
      tags:
      - "coach"
      summary: "Update service for coach"
      operationId: "coachUpdate"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        description: "token of partner"
        required: true
        type: "string"
      - name: "param"
        in: "path"
        description: "parameter filter"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "data credential guest"
        required: true
        schema:
          $ref: "#/definitions/coachUpdateParam"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/notifSuccess"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "coach"
    get:
      tags:
      - "coach"
      summary: "Add new place off partner"
      operationId: "coachList"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        description: "token of partner"
        required: true
        type: "string"
      - name: "param"
        in: "path"
        description: "parameter filter"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/notifSuccess"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "coach"
...
  /member/join:
    post:
      tags:
      - "member"
      summary: "join membership"
      description: ""
      operationId: "joinMember"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data credential guest"
        required: true
        schema:
          $ref: "#/definitions/joinMemberParam"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/apiResponse"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "member"
  /member/fee/{token}:
    get:
      tags:
      - "member"
      summary: "fee of membership"
      description: ""
      operationId: "memberFee"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        description: "token of member device"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/apiResponse"
        400:
          description: "Invalid status value"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
      x-swagger-router-controller: "member"
  /member/act:
    # get:
    #   tags:
    #   - "member"
    #   summary: "fee of membership"
    #   description: ""
    #   operationId: "memberClassBooked"
    #   consumes:
    #   - "multipart/form-data"
    #   - "application/x-www-form-urlencoded"
    #   - "application/json"
    #   produces:
    #   - "application/xml"
    #   - "application/json"
    #   parameters:
    #   - name: "token"
    #     in: "header"
    #     description: "token of member device"
    #     required: true
    #     type: "string"
    #   responses:
    #     200:
    #       description: "Success"
    #     500:
    #       description: "Internal error"
    #     401:
    #       description: "Unauthorize"
    #   security:
    #   - petstore_auth:
    #     - "write:pets"
    #     - "read:pets"
    #   x-swagger-router-controller: "member"

    post:
      tags:
      - "member"
      summary: "fee of membership"
      description: ""
      operationId: "memberActivity"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data credential guest"
        required: true
        schema:
          $ref: "#/definitions/memberActivityParam"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
      x-swagger-router-controller: "member"
  /member/personalRecord/category/{token}:
    get:
      tags:
      - "member"
      summary: "Personal record"
      description: ""
      operationId: "getPersonalRecordCategory"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        description: "token of member device"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
      x-swagger-router-controller: "member"
  /member/personalRecord/{token}:
    get:
      tags:
      - "member"
      summary: "Personal record"
      description: ""
      operationId: "getPersonalRecord"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        description: "token of member device"
        required: true
        type: "string"
      - name: "param" # all / {"prCat":""}
        in: "header"
        description: "token of member device"
        required: true
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
      x-swagger-router-controller: "member"

    post:
      tags:
      - "member"
      summary: "Personal record"
      description: ""
      operationId: "personalRecord"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        description: "token of member device"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "data credential guest"
        required: true
        schema:
          $ref: "#/definitions/personalRecordParam"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
      x-swagger-router-controller: "member"
  /coach/class/schedule/{token}:
    post:
      tags:
      - "coach"
      summary: "Create Schedule access"
      description: ""
      operationId: "createSchedule" 
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        description: "token of member device"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "data credential guest"
        required: true
        schema:
          $ref: "#/definitions/createSchedule"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "coach"
    get:
      tags:
      - "coach"
      summary: "Get Schedule access"
      description: ""
      operationId: "getSchedule" 
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      - "application/json"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "token"
        in: "path"
        description: "token of member device"
        required: true
        type: "string"
      - name: "filter"
        in: "header"
        description: "filter for class schedule"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
        500:
          description: "Internal error"
        401:
          description: "Unauthorize"
        404:
          description: "Data not found"
      x-swagger-router-controller: "coach"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  coachUpdateParam:
    type: "object"
    properties:
      idSelfSchedule:
        type: "integer"
        format: "int32"
      selfId:
        type: "integer"
        format: "int32"
      targetSchedule:
        type: "integer"
        format: "int32"
      targetCoach:
        type: "integer"
        format: "int32"
...
  joinMemberParam:
    type: "object"
    properties:
      token:
        type: "string"
      memberCat:
        type: "string"
...
  memberActivityParam:
    type: "object"
    properties:
      token:
        type: "string"
      scheduleId:
        # type: "integer"
        # format: "int32"
        type: "string"
      action:
        type: "string"
...
  personalRecordParam:
    type: "object"
    properties:
      # token:
      #   type: "string"
      prCat:
        type: "string"
      dataValue:
        type: "string"
  createSchedule:
    type: "object"
    properties:
      classId:
        type: "integer"
      startDate:
        type: "string"
      endDate:
        type: "string"
      startTime:
        type: "string"
      endTime:
        type: "string"
  notifSuccess:
    type: "object"
    properties:
      message:
        type: "string"
      data:
        type: "string"
    example:
      message: "Success"
      data: "data all customer"
  apiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
      data:
        type: "object"
        properties: {}
    example:
      code: 0
      type: "type"
      message: "message"
      data:
        token: "ABCDEFGHIJKLMNOPQ2344222"
...